+++ autosub/__init__.py
@@ -137,18 +137,20 @@ def main():  # pylint: disable=too-many-branches, too-many-statements, too-many-
                         "[sample_rate]",
                         "{sample_rate}".format(sample_rate=args.api_sample_rate))
 
-                if args.api_suffix == ".ogg":
-                    # regard ogg as ogg_opus
-                    args.audio_split_cmd = \
-                        args.audio_split_cmd.replace(
-                            "-vn",
-                            "-vn -c:a libopus")
-                elif args.api_suffix == ".pcm":
-                    # raw pcm
-                    args.audio_split_cmd = \
-                        args.audio_split_cmd.replace(
-                            "-vn",
-                            "-vn -c:a pcm_s16le -f s16le")
+                if args.audio_split_cmd == constants.DEFAULT_AUDIO_SPLT:
+                    # if user doesn't modify the audio_split_cmd
+                    if args.api_suffix == ".ogg":
+                        # regard ogg as ogg_opus
+                        args.audio_split_cmd = \
+                            args.audio_split_cmd.replace(
+                                "-vn",
+                                "-vn -c:a libopus")
+                    elif args.api_suffix == ".pcm":
+                        # raw pcm
+                        args.audio_split_cmd = \
+                            args.audio_split_cmd.replace(
+                                "-vn",
+                                "-vn -c:a pcm_s16le -f s16le")
 
             cmdline_utils.validate_aovp_args(args)
             fps = cmdline_utils.get_fps(args=args, input_m=input_m)